requirements.txt
torch==2.2.0
torchvision==0.17.0
transformers==4.37.2
soundfile==0.13.1
pandas==2.1.3
numpy==1.26.4
sentencepiece==0.2.0
protobuf==3.19.6
pydub==0.25.1
simpleaudio==1.0.4
openai==1.61.1
python-dotenv==1.0.1
mlflow==2.20.2
librosa==0.10.2
datasets==3.3.1
elevenlabs==1.51.0
accelerate==0.28.0
num2words==0.5.14